+++ common/src/main/java/mca/Config.java
@@ -59,6 +59,7 @@ public final class Config implements Serializable {
     public int interactionFatigueCooldown = 4800;
     public float taxesFactor = 0.5f;
     public boolean enterVillageNotification = true;
+    public boolean showNameTags = true;
 
     public static File getConfigFile() {
         return new File("./config/mca.json");