+++ changelog.md
@@ -184,3 +184,4 @@
 * Staff of Life can no longer be enchanted
 * Fixed chores phrase names
 * Command kill no longer counts as murder
+* Added config flag to disable name tags